Move trigger source to own file
Having such a large nested class inside a small top level class is VERY
confusing.
